Move tests to be colocated with src
