+++ typescript/sdk/package.json
@@ -45,8 +45,8 @@
     "prepublishOnly": "yarn build",
     "prettier": "prettier --write ./src ./test",
     "test": "yarn test:unit && yarn test:hardhat",
-    "test:unit": "mocha --config .mocharc.json './test/{,!(hardhat)/**}/*.test.ts'",
-    "test:hardhat": "hardhat test ./test/hardhat/*.ts"
+    "test:unit": "mocha --config .mocharc.json './src/{,!(hardhat)/**}/*.test.ts'",
+    "test:hardhat": "hardhat test ./src/hardhat/*.test.ts"
   },
   "types": "dist/index.d.ts"
 }
